Use git submodule for rapid check
Prior to this commit this repo used a static library and the header files for
rapid check rather than a git submodule to work with ROS build system. This
commit removes the duplicate headers and static library in favor of the
submodule option while still working with the ros test process.