Update Documentation
Prior to this commit the documentation was lacking in clarity of what
dependencies were required prior to using the components in ROSCCO. Also the
README was lacking in clear header organizatoin. This commit fixes that by
reorganizing the README headers, providing more details about the dependencies
for this project and providing clarity of the Licensing information as well as
corrections to the grammer in the description of the package.