Fix CLI steering command doc bug
Prior to this commit the documentation in this project's README.md had a typo in it's example for sending steering torque commands on the command line. This resulted in a failure to actuate steering with the example command using these docs as your guide. After this commit one reference to `SteeringCommand` has been revised to `steering_command`, the fix required for the command to succeed.