Increase sleep time for build delay
Prior to this commit tests were intermittently failing due to a backlog in the
message passing on slower build machines inside docker containers. This commit
fixes that by increasing the sleep time by 2ms so that the test has enough wait
time for messages to be passed across the ROS bus in the slower environment.